Fix CI: backend lint config and frontend lockfile/peer-dep drift
Backend (ruff was failing on default rule set, never configured):
- Add pyproject.toml with a FastAPI-appropriate ruff config (ignore
  B008 for Depends() defaults, BLE001 for intentional broad excepts
  at service/worker boundaries, UP042 to keep str+Enum for
  SQLAlchemy/Pydantic compat)
- Fix the real issues ruff found: unused imports, unsorted __all__,
  datetime.utcnow() -> timezone-aware utcnow(), redundant int(round()),
  sorted()[0] -> min(), an unused variable, a stale noqa comment

Frontend (npm ci was failing in CI):
- The lockfile was missing some transitive optional-dependency entries
  (@emnapi/*) because it grew incrementally across many `npm install`/
  `shadcn add` calls instead of one clean resolve, and was generated
  with npm 11 while CI's Node 22 runner ships npm 10 - regenerated it
  cleanly with npm 10 to match
- That surfaced a real react-hook-form/@hookform-resolvers peer range
  mismatch; added .npmrc with legacy-peer-deps=true
- Also fixes a real bug this exposed: AuthProvider read the auth token
  in a useEffect instead of a lazy useState initializer, so a hard
  page load briefly saw "no token" and AuthGuard would redirect an
  already-logged-in user to /login

+++ backend/app/benchmark/engine.py
@@ -29,7 +29,7 @@ def timed_inference_stats(run_once: Callable[[], None], iterations: int = 50, wa
         samples_ms.append((time.perf_counter() - start) * 1000.0)
 
     samples_ms.sort()
-    p95_index = min(len(samples_ms) - 1, int(round(0.95 * (len(samples_ms) - 1))))
+    p95_index = min(len(samples_ms) - 1, round(0.95 * (len(samples_ms) - 1)))
     return {
         "mean_ms": statistics.fmean(samples_ms),
         "median_ms": statistics.median(samples_ms),
